+++ app/src/main/AndroidManifest.xml
@@ -15,12 +15,12 @@
         android:theme="@style/Theme.GangwonHyuil"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.OnBodingActivity"
+            android:name=".ui.onboarding.OnBodingActivity"
             android:exported="true" >
 
         </activity>
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
