--- app/src/main/java/com/gangwonhyuil/gangwonhyuil/data/repositoryImpl/KakaoLoginRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package com.gangwonhyuil.gangwonhyuil.data.repositoryImpl
-
-class KakaoLoginRepositoryImpl {
-}
