--- app/src/main/java/com/gangwonhyuil/gangwonhyuil/data/repositoryImpl/PostRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package com.gangwonhyuil.gangwonhyuil.data.repositoryImpl
-
-class PostRepositoryImpl {
-}
