--- app/src/main/java/com/gangwonhyuil/gangwonhyuil/domain/entity/ReviewEntity.kt
@@ -1,4 +1,0 @@
-package com.gangwonhyuil.gangwonhyuil.domain.entity
-
-class ReviewEntity {
-}
