--- app/src/main/java/com/gangwonhyuil/gangwonhyuil/domain/entity/TourEntity.kt
@@ -1,4 +1,0 @@
-package com.gangwonhyuil.gangwonhyuil.domain.entity
-
-class TourEntity {
-}
