--- app/src/main/java/com/gangwonhyuil/gangwonhyuil/domain/repository/OfficeRepository.kt
@@ -1,4 +1,0 @@
-package com.gangwonhyuil.gangwonhyuil.domain.repository
-
-class OfficeRepository {
-}
