--- app/src/main/java/com/gangwonhyuil/gangwonhyuil/domain/repository/PostRepository.kt
@@ -1,4 +1,0 @@
-package com.gangwonhyuil.gangwonhyuil.domain.repository
-
-class PostRepository {
-}
