--- app/src/main/java/com/gangwonhyuil/gangwonhyuil/domain/repository/TourRepository.kt
@@ -1,4 +1,0 @@
-package com.gangwonhyuil.gangwonhyuil.domain.repository
-
-class TourRepository {
-}
