+++ app/src/main/java/com/gangwonhyuil/gangwonhyuil/ui/community/CommunityFragment.kt
@@ -1,4 +1,4 @@
-package com.gangwonhyuil.gangwonhyuil.presentation
+package com.gangwonhyuil.gangwonhyuil.ui.community
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [AiFragment.newInstance] factory method to
+ * Use the [CommunityFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class AiFragment : Fragment() {
+class CommunityFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class AiFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_ai, container, false)
+        return inflater.inflate(R.layout.fragment_community, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class AiFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment AiFragment.
+         * @return A new instance of fragment ComunityFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            AiFragment().apply {
+            CommunityFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
