+++ app/src/main/res/layout/activity_on_boding.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.OnBodingActivity">
+    tools:context=".ui.onboarding.OnBodingActivity">
 
     <TextView
         android:id="@+id/tv_result"
