+++ app/src/main/res/layout/fragment_ai.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.AiFragment">
+    tools:context=".ui.ai.AiFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
