+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.gangwonhyuil.gangwonhyuil.presentation.HomeFragment"
+        android:name="com.gangwonhyuil.gangwonhyuil.ui.home.HomeFragment"
         android:label="HomeFragment"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/aiFragment"
-        android:name="com.gangwonhyuil.gangwonhyuil.presentation.AiFragment"
+        android:name="com.gangwonhyuil.gangwonhyuil.ui.ai.AiFragment"
         android:label="BuildingFragment"
         tools:layout="@layout/fragment_ai" />
 
     <fragment
         android:id="@+id/communityFragment"
-        android:name="com.gangwonhyuil.gangwonhyuil.presentation.CommunityFragment"
+        android:name="com.gangwonhyuil.gangwonhyuil.ui.community.CommunityFragment"
         android:label="FacilityFragment"
         tools:layout="@layout/fragment_community" />
 
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.gangwonhyuil.gangwonhyuil.presentation.ProfileFragment"
+        android:name="com.gangwonhyuil.gangwonhyuil.ui.profile.ProfileFragment"
         android:label="NoticeFragment"
         tools:layout="@layout/fragment_profile" />
